Move stuff to primitive
