+++ src/compiler/typescript_to_rust/azle_generate/src/candid_type/primitive/mod.rs
@@ -1,35 +1,74 @@
 use cdk_framework::act::node::{candid::Primitive, CandidType};
-use swc_ecma_ast::TsTypeRef;
+use swc_common::Span;
+use swc_ecma_ast::{TsKeywordType, TsKeywordTypeKind};
 
-use crate::{traits::GetName, ts_ast::SourceMapped};
+use crate::{
+    errors::{ErrorMessage, Suggestion},
+    traits::{GetSourceFileInfo, GetSourceInfo, GetSourceText, GetSpan},
+    ts_ast::SourceMapped,
+};
 
-impl SourceMapped<'_, TsTypeRef> {
+impl SourceMapped<'_, TsKeywordType> {
     pub fn to_candid_type(&self) -> CandidType {
-        match self.get_name() {
-            "blob" => CandidType::Primitive(Primitive::Blob),
-            "float32" => CandidType::Primitive(Primitive::Float32),
-            "float64" => CandidType::Primitive(Primitive::Float64),
-            "int" => CandidType::Primitive(Primitive::Int),
-            "int8" => CandidType::Primitive(Primitive::Int8),
-            "int16" => CandidType::Primitive(Primitive::Int16),
-            "int32" => CandidType::Primitive(Primitive::Int32),
-            "int64" => CandidType::Primitive(Primitive::Int64),
-            "nat" => CandidType::Primitive(Primitive::Nat),
-            "nat8" => CandidType::Primitive(Primitive::Nat8),
-            "nat16" => CandidType::Primitive(Primitive::Nat16),
-            "nat32" => CandidType::Primitive(Primitive::Nat32),
-            "nat64" => CandidType::Primitive(Primitive::Nat64),
-            "Principal" => CandidType::Primitive(Primitive::Principal),
-            "empty" => CandidType::Primitive(Primitive::Empty),
-            "reserved" => CandidType::Primitive(Primitive::Reserved),
-            "text" => CandidType::Primitive(Primitive::String),
-            "Opt" => CandidType::Opt(self.to_option()),
-            "Func" => CandidType::Func(self.to_func(None)),
-            "Record" => CandidType::Record(self.to_record()),
-            "Tuple" => CandidType::Tuple(self.to_tuple()),
-            "Vec" => CandidType::Array(self.to_vec()),
-            "Variant" => CandidType::Variant(self.to_variant()),
-            _ => CandidType::TypeRef(self.to_type_ref()),
+        match self.kind {
+            TsKeywordTypeKind::TsBooleanKeyword => CandidType::Primitive(Primitive::Bool),
+            TsKeywordTypeKind::TsStringKeyword => CandidType::Primitive(Primitive::String),
+            TsKeywordTypeKind::TsVoidKeyword => CandidType::Primitive(Primitive::Void),
+            TsKeywordTypeKind::TsNullKeyword => CandidType::Primitive(Primitive::Null),
+            TsKeywordTypeKind::TsNumberKeyword => CandidType::Primitive(Primitive::Float64),
+            _ => panic!("{}", self.unsupported_type_error()),
         }
+    }
+
+    pub(super) fn unsupported_type_error(&self) -> ErrorMessage {
+        match &self.kind {
+            TsKeywordTypeKind::TsBigIntKeyword => self.bigint_not_supported_error(),
+            TsKeywordTypeKind::TsObjectKeyword => self.keyword_not_supported_error(),
+            TsKeywordTypeKind::TsNeverKeyword => self.keyword_not_supported_error(),
+            TsKeywordTypeKind::TsSymbolKeyword => self.keyword_not_supported_error(),
+            TsKeywordTypeKind::TsIntrinsicKeyword => self.keyword_not_supported_error(),
+            TsKeywordTypeKind::TsUndefinedKeyword => self.keyword_not_supported_error(),
+            TsKeywordTypeKind::TsUnknownKeyword => self.keyword_not_supported_error(),
+            TsKeywordTypeKind::TsAnyKeyword => self.keyword_not_supported_error(),
+            _ => panic!("Unreachable: {} is supported", self.get_source_text()),
+        }
+    }
+
+    fn bigint_not_supported_error(&self) -> ErrorMessage {
+        let replacement = "int".to_string();
+        let suggestion = Some(Suggestion {
+            title: "`int` will cover most everything that `bigint` does. For more number type options see: https://internetcomputer.org/docs/current/references/candid-ref/#type-nat".to_string(),
+            range: self.source_map.generate_modified_range(self.span, &replacement),
+            source: self.source_map.generate_modified_source(self.span, &replacement),
+            annotation: Some("Try using `int` here.".to_string()),
+            import_suggestion: Some("import { int } from 'azle';".to_string()),
+        });
+        ErrorMessage {
+            title: "Unsupported Type".to_string(),
+            origin: self.get_origin(),
+            line_number: self.get_line_number(),
+            source: self.get_source(),
+            range: self.get_range(),
+            annotation: "bigint is not a supported type".to_string(),
+            suggestion,
+        }
+    }
+
+    fn keyword_not_supported_error(&self) -> ErrorMessage {
+        ErrorMessage {
+            title: "Unsupported Type".to_string(),
+            origin: self.get_origin(),
+            line_number: self.get_line_number(),
+            source: self.get_source(),
+            range: self.get_range(),
+            annotation: format!("{} is not a supported type", self.get_source_text()),
+            suggestion: None,
+        }
+    }
+}
+
+impl GetSpan for TsKeywordType {
+    fn get_span(&self) -> Span {
+        self.span
     }
 }
