+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_type/mod.rs
@@ -6,7 +6,6 @@ use crate::ts_ast::SourceMapped;
 mod get_span;
 mod to_candid_type;
 mod ts_fn_or_constructor_type;
-mod ts_keyword_type;
 mod ts_type_lit;
 mod ts_type_ref;
 mod type_to_string;
