--- src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_type/ts_keyword_type/mod.rs
@@ -1,3 +1,0 @@
-mod errors;
-mod get_span;
-mod to_candid_type;
