+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_type/ts_type_ref/mod.rs
@@ -7,6 +7,7 @@ use crate::ts_ast::SourceMapped;
 mod errors;
 mod get_name;
 mod get_span;
+mod to_candid_type;
 
 impl SourceMapped<'_, TsTypeRef> {
     pub fn get_ts_type(&self) -> SourceMapped<TsType> {
